Create legacy screen
