+++ magellan-library/src/main/java/com/wealthfront/magellan/ScreenView.java
@@ -1,5 +1,7 @@
 package com.wealthfront.magellan;
 
+import com.wealthfront.magellan.flow.Screen;
+
 /**
  * The interface to be implemented by a View representing a Screen.
  * @see BaseScreenView
